remove the LinkToolbarTransientMenu class/constructor thingy. (yippee!)
this class was used when we encountered an unknown rel type on a <html:link/>.  it generated a <menu> and <menupopup> the first time getXULElement and getPopup were called.  we now generate this in linkToolbarHandler.createItemForLinkType instead, at the point when we first handle the link with the unrecognized rel.  hence the class is pointless

having done this, getXULElement can probably be eliminated in favour of the xulElement property, which we can initialise in the constructor.