rewrite ltLanguageDictionary to not use srGetStrBundle because the file it is defined in, strres.js, is probably going to be removed from Firebird soon.
implement caching of the linkInfo objects produced by linkToolbarHandler on the document of the relevant link element.  these can then be retrieved on tab switching, rather than calling fullSlowRefresh.

Note: when enabling the link toolbar we need to call fullSlowRefresh, but this patch does not do that

also: interaction with the ltUI.(de)activate() functions may be a bit weird.  need to do more testing