mostly fix mozbug 105109 - better handling of unrecognised rel types.
a new item class LinkToolbarTransientItem has been added, which creates itself a <menuitem/> and a <menu/> element, and shows/hides the two as required (i.e when it goes from handling 1 to 2 links).

at the moment the menuitem always uses the rel type as its label, but using the title when available might be better. (requires a rethink of the linkInfo objects though)

this also includes a fix for groups of links with an unrecognised rel but no title, for which the menuitems previously had no label.  they now use the url


the new class does not inherit from any of the existing ones, which is something to fix later.