in recent nighlies of Firebird (20030915 and 20030916 tested), the bindings are for the toolbox elements ar ebeing attached in the opposite order to preiviously.  this means that the linked-toolboxes were being constucted before the navigator-toolbox, so it's |palette| field wasn't ready for them to use.  the result was that the custom toolbars in other toolboxes ended up empty when rebuilt on startup
the solution is to do away with linked toolboxes, and have toolboxes all use a "palette" attribute, which takes the id of the <toolbarpalette/> they wabnt to use.  the first toolbox constructed will then remove the palette from the document and store it in a hashtable-by-id called |toolbarpalettes| as a property of the |document|.

will most likely not work with older builds, because there the toolbar elements are constructed before the toolboxes, and so finding the paletet was done in the toolbar binding.  in new builds toolboxes get constructed before their toolbar children, so there is no need for such hackery